extras/yt_video_transcript_api/yt_transcript/packages/academi/yt_transcript/requirements.txt
youtube_transcript_api
requests
yt_dlp
deepgram-sdk==3.5.2
pydantic
python-dotenv